[Refactor] Move BaseReader and its child class to FileReader.cpp/.h
