+++ cpp-ch/local-engine/tests/utils/ReaderTestBase.h
@@ -41,46 +41,6 @@ class Block;
 namespace local_engine::test
 {
 
-
-// TODO: CppToDataType move to other cpp files
-template <typename T> struct CppToDataType;
-
-template <>
-struct CppToDataType<Int64>
-{
-    using Type = DB::DataTypeInt64;
-    using ColumnType = DB::ColumnInt64;
-    static auto create() { return std::make_shared<Type>(); }
-};
-
-template <>
-struct CppToDataType<UInt64>
-{
-    using Type = DB::DataTypeUInt64;
-    using ColumnType = DB::ColumnUInt64;
-    static auto create() { return std::make_shared<Type>(); }
-};
-
-template <typename T >
-DB::ColumnPtr makeColumn(const std::vector<T>& data)
-requires (std::is_base_of_v<DB::ColumnVector<T>, typename CppToDataType<T>::ColumnType>)
-{
-    static_assert(!DB::is_decimal<T>);
-
-    auto column = CppToDataType<T>::ColumnType::create(data.size());
-    typename DB::ColumnVector<T>::Container & vec = column->getData();
-    memcpy(vec.data(), data.data(), data.size() * sizeof(T));
-    return column;
-}
-
-template <typename T>
-DB::ColumnWithTypeAndName makeColumn(const std::vector<T>& data, const std::string & col_name)
-requires (std::is_base_of_v<DB::ColumnVector<T>, typename CppToDataType<T>::ColumnType>)
-{
-    return {makeColumn(data), CppToDataType<T>::create(), col_name};
-}
-// end of CppToDataType
-
 template <typename T>
 concept couldbe_collected = requires(T t)
 {
